feat: consolidate outcome evaluation script
