+++ README.md
@@ -1,21 +1,20 @@
 # SP500SCN
 
-## Hit Checking
+## Outcome Evaluation
 
-Use the `scripts/check_hits.py` utility to update `data/history/outcomes.csv` with trade results.
+Use the `scripts/evaluate_outcomes.py` utility to update `data/history/outcomes.csv` with trade results.
 
-```bash
-python scripts/check_hits.py
-```
-
-The script fetches daily price data for each pending entry and marks hits or misses accordingly.
-
-## Historical Scoring
-
-For backtesting datasets that include `EvalDate`, `WindowEnd`, and `TargetLevel`
-columns, run `scripts/score_history.py` to evaluate whether price targets were
-hit before each window closed.
+For pending trades:
 
 ```bash
-python scripts/score_history.py
+python scripts/evaluate_outcomes.py --mode pending
 ```
+
+For historical datasets that include `EvalDate`, `WindowEnd`, and `TargetLevel` columns:
+
+```bash
+python scripts/evaluate_outcomes.py --mode historical
+```
+
+The script fetches daily price data for each entry and marks hits or misses accordingly.
+
